code cleanup and type aliasing

+++ libgramtools/include/masks.hpp
@@ -13,7 +13,7 @@ public:
     std::vector<uint64_t> sites;
 
     std::vector<int> allele;
-    std::vector<std::vector<float>> allele_coverage;
+    std::vector<std::vector<double>> allele_coverage;
 
     MasksParser(){};
     MasksParser(const std::string &sites_fname, const std::string &alleles_fname);
